20161229 - more can checks on persons show blade for relationships

+++ resources/views/persons/show.blade.php
@@ -200,31 +200,35 @@
                     <div class='panel-heading'>
                         <h2><strong>Relationships</strong></h2>
                     </div>
+                    @can('create-relationship')
                     {!! Form::open(['method' => 'POST', 'route' => ['relationship_type.addme']]) !!}
                     {!! Form::label('relationship_type', 'Add Relationship: ', ['class' => 'col-md-2'])  !!}
                     {!! Form::select('relationship_type', $relationship_types, NULL, ['class' => 'col-md-2']) !!}
                     {!! Form::hidden('contact_id',$person->id)!!}
                     {!! Form::submit('Create relationship') !!}
                     {!! Form::close() !!}
-                    
+                    @endCan
                             <ul>    
                                 @foreach($person->a_relationships as $a_relationship)
                                 <li>
-                                    {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $a_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
-                                    {!!$person->contact_link!!} is {{ $a_relationship->relationship_type->label_a_b }} {!! $a_relationship->contact_b->contact_link !!} 
-                                    
-                                    {!! Form::image('img/delete.png','btnDelete',['title'=>'Delete Relationship '.$a_relationship->id, 'style'=>'padding-left: 50px;']) !!} 
-                                    {!! Form::close() !!}
+                                    {!!$person->contact_link_full_name!!} is {{ $a_relationship->relationship_type->label_a_b }} {!! $a_relationship->contact_b->contact_link_full_name !!} 
+                                    @can('delete-relationship')
+                                        {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $a_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
+                                        {!! Form::image('img/delete.png','btnDelete',['title'=>'Delete Relationship '.$a_relationship->id, 'style'=>'padding-left: 50px;']) !!} 
+                                        {!! Form::close() !!}
+                                    @endCan
                                 </li>
                                 @endforeach
                                 
                    
                                 @foreach($person->b_relationships as $b_relationship)
-                                <li>
-                                    {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $b_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
-                                        {!!$person->contact_link!!} is {{ $b_relationship->relationship_type->label_b_a }} {!! $b_relationship->contact_a->contact_link!!}
+                                <li>    
+                                    {!!$person->contact_link_full_name!!} is {{ $b_relationship->relationship_type->label_b_a }} {!! $b_relationship->contact_a->contact_link_full_name!!}
+                                    @can('delete-relationship')    
+                                        {!! Form::open(['method' => 'DELETE', 'route' => ['relationship.destroy', $b_relationship->id],'onsubmit'=>'return ConfirmDelete()']) !!}
                                         {!! Form::image('img/delete.png','btnDelete',['title'=>'Delete Relationship '.$b_relationship->id,'style'=>'padding-left: 50px;']) !!} 
-                                    {!! Form::close() !!}
+                                        {!! Form::close() !!}
+                                    @endCan
                                 </li>
                                 @endforeach
